Mudando tem min e max só para temp
agora é só temperatura

+++ app/src/main/java/br/com/orlando/weatherforecast/returnApi/Main.kt
@@ -5,4 +5,4 @@ data class Main(val temp: Double = 0.0,
                 val humidity: Int = 0,
                 val pressure: Int = 0,
                 val feelsLike: Double = 0.0,
-                val tempMax: Int = 0)
+                val tempMax: Double = 0.0)
